refactor: codeblock header now rendered by multicodeblock
